requirement.txt
Django==2.0.1
django-ckeditor==5.6.1
django-el-pagination==3.2.4
django-js-asset==1.1.0
django-registration-redux==2.5
django-widget-tweaks==1.4.3
Pillow==5.3.0
psycopg2==2.7.6.1
pytz==2018.7
